(test) test truck menu without requiring redux
I believe this is the proper way to test redux components - split the component that simply expects props, and the container that connects that component to the store. For now we can keep that all in one file and just add an export. this allows us to test functionality without configuring the store for every test.